Refactor(auth): Konsolidasi rute reset kata sandi untuk memperbaiki bug 404
Memindahkan logika reset kata sandi dari file terpisah ke dalam rute auth utama.

Sebelumnya, rute reset kata sandi berada di file mereka sendiri dan dimuat di bawah /api/passwordReset, tetapi frontend salah mencoba mengaksesnya melalui /api/auth. Hal ini menyebabkan kesalahan 404.

Perubahan ini mengkonsolidasikan rute-rute tersebut ke dalam file rute auth utama dan memperbarui layanan frontend untuk menunjuk ke endpoint yang benar (/api/auth/forgot-password, dll.). Ini menyelesaikan bug 404 dan menyederhanakan struktur rute backend.
