--- backend/routes/passwordReset.js
@@ -1,157 +1,0 @@
-const express = require('express');
-const pool = require('../db/connection');
-const bcrypt = require('bcrypt');
-const { v4: uuidv4 } = require('uuid');
-const emailService = require('../services/emailService');
-const dayjs = require('dayjs');
-
-const router = express.Router();
-
-// -----------------------------
-// Request reset password
-// -----------------------------
-router.post('/forgot-password', async (req, res) => {
-  const { email } = req.body;
-  if (!email) {
-    return res.status(400).json({ success: false, message: 'Email is required' });
-  }
-
-  let connection;
-  try {
-    connection = await pool.getConnection();
-    const [rows] = await connection.query('SELECT * FROM users WHERE email = ?', [email]);
-    const user = rows[0];
-
-    if (!user) {
-      return res.json({ success: true, message: 'Jika email terdaftar, instruksi reset password telah dikirim' });
-    }
-
-    if (user.reset_attempts >= 10) {
-      const lastAttempt = dayjs(user.updated_at);
-      if (lastAttempt.isAfter(dayjs().subtract(1, 'hour'))) {
-        return res.status(429).json({ success: false, message: 'Terlalu banyak percobaan reset. Coba lagi dalam 1 jam' });
-      }
-    }
-
-    const resetToken = uuidv4();
-    console.log(`[PasswordReset] Generated token for ${email}: ${resetToken}`);
-
-    // Simpan expired sesuai timezone server (bukan UTC)
-    const expiresAt = dayjs().add(1, 'hour').format('YYYY-MM-DD HH:mm:ss');
-    const attempts = user.reset_attempts >= 10 ? 1 : (user.reset_attempts || 0) + 1;
-
-    const [updateResult] = await connection.query(
-      `UPDATE users 
-       SET reset_token = ?, reset_token_expires_at = ?, reset_attempts = ?, updated_at = NOW() 
-       WHERE email = ?`,
-      [resetToken, expiresAt, attempts, email]
-    );
-
-    console.log(`[PasswordReset] Database update result for ${email}: affectedRows = ${updateResult.affectedRows}`);
-
-    const emailSent = await emailService.sendPasswordResetEmail(email, resetToken, user.username);
-    if (emailSent) {
-      res.json({ success: true, message: 'Instruksi reset password telah dikirim ke email Anda' });
-    } else {
-      res.status(500).json({ success: false, message: 'Gagal mengirim email reset password' });
-    }
-  } catch (error) {
-    console.error('Error in forgot password:', error);
-    res.status(500).json({ success: false, message: 'Terjadi kesalahan server' });
-  } finally {
-    if (connection) connection.release();
-  }
-});
-
-// -----------------------------
-// Verify reset token
-// -----------------------------
-router.get('/verify-reset-token', async (req, res) => {
-  const { token } = req.query;
-  console.log(`[PasswordReset] Received token for verification: ${token}`);
-
-  if (!token) {
-    return res.status(400).json({ success: false, message: 'Token is required' });
-  }
-
-  let connection;
-  try {
-    connection = await pool.getConnection();
-    const [rows] = await connection.query('SELECT * FROM users WHERE reset_token = ?', [token]);
-    const user = rows[0];
-
-    console.log('[PasswordReset] Query result for token:', rows);
-
-    if (!user || !user.reset_token_expires_at) {
-      return res.status(400).json({ success: false, message: 'Token tidak valid' });
-    }
-
-    if (dayjs().isAfter(dayjs(user.reset_token_expires_at))) {
-      return res.status(400).json({ success: false, message: 'Token telah kedaluwarsa' });
-    }
-
-    res.json({ success: true, message: 'Token valid', email: user.email });
-  } catch (error) {
-    console.error('Error verifying reset token:', error);
-    res.status(500).json({ success: false, message: 'Terjadi kesalahan server' });
-  } finally {
-    if (connection) connection.release();
-  }
-});
-
-// -----------------------------
-// Reset password
-// -----------------------------
-router.post('/reset-password', async (req, res) => {
-  const { token, password, confirmPassword } = req.body;
-
-  if (!token || !password || !confirmPassword) {
-    return res.status(400).json({ success: false, message: 'Semua field harus diisi' });
-  }
-  if (password !== confirmPassword) {
-    return res.status(400).json({ success: false, message: 'Konfirmasi password tidak cocok' });
-  }
-  if (password.length < 6) {
-    return res.status(400).json({ success: false, message: 'Password minimal 6 karakter' });
-  }
-
-  let connection;
-  try {
-    connection = await pool.getConnection();
-    await connection.beginTransaction();
-
-    const [rows] = await connection.query('SELECT * FROM users WHERE reset_token = ? FOR UPDATE', [token]);
-    const user = rows[0];
-
-    console.log('[PasswordReset] Reset password query result:', rows);
-
-    if (!user || !user.reset_token_expires_at) {
-      await connection.rollback();
-      return res.status(400).json({ success: false, message: 'Token tidak valid' });
-    }
-
-    if (dayjs().isAfter(dayjs(user.reset_token_expires_at))) {
-      await connection.rollback();
-      return res.status(400).json({ success: false, message: 'Token telah kedaluwarsa' });
-    }
-
-    const hashedPassword = await bcrypt.hash(password, 12);
-    await connection.query(
-      `UPDATE users 
-       SET password_hash = ?, reset_token = NULL, reset_token_expires_at = NULL, reset_attempts = 0, updated_at = NOW() 
-       WHERE id = ?`,
-      [hashedPassword, user.id]
-    );
-
-    await connection.commit();
-    res.json({ success: true, message: 'Password berhasil direset. Silakan login dengan password baru' });
-  } catch (error) {
-    if (connection) await connection.rollback();
-    console.error('Error resetting password:', error);
-    res.status(500).json({ success: false, message: 'Terjadi kesalahan server' });
-  } finally {
-    if (connection) connection.release();
-  }
-});
-
-module.exports = router;
